engine/requirements.txt
openai-whisper>=20231117
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
Pillow>=10.0.0
openai>=1.0.0
anthropic>=0.25.0
httpx>=0.25.0
numpy>=1.24.0
gguf>=0.10.0
diffusers>=0.30.0
soundfile>=0.12.0
torch>=2.5.0
transformers>=4.40.0
voxcpm
